% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{rx_calc}
\alias{rx_calc}
\title{Create a reactive calculation (reactive)}
\usage{
rx_calc(x, env = parent.frame(), quoted = FALSE, label = NULL,
  domain = getDefaultReactiveDomain(), ..stacktraceon = TRUE)
}
\description{
Currently just an alias for \code{\link[shiny]{reactive}}.
}

