% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rx_calc_with_previous}
\alias{rx_calc_with_previous}
\title{Reactive calculation with previous value}
\usage{
rx_calc_with_previous(func, initialValue = NULL)
}
\arguments{
\item{func}{A function that takes a single parameter. Each time this function
is called, it will be passed its previous return value.}

\item{initialValue}{The value to pass to the function the first time it is
called.}
}
\value{
An \code{\link{rx_calc}} based on the function.
}
\description{
Use this variant of \code{\link{rx_calc}} when your logic depends on the
previous value you generated. For example, you might want to accumulate some
state over each invalidation; or in some exceptional circumstances you may
want to return the previous value itself with no changes.
}

