% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{rx_domain}
\alias{rx_domain}
\title{Create or run within a reactive domain}
\usage{
rx_domain(expr)
}
\arguments{
\item{expr}{The expression or code block to evaluate under the domain. If
provided, the domain will automatically be closed before returning.}
}
\value{
If \code{expr} is missing, a domain object; call the \code{close()}
  method on it to terminate the domain and clean up registered objects. If
  \code{expr} is provided, then the value of \code{expr} is returned.
}
\description{
A reactive domain is an environment within which reactive constructs can run.
Some objects like \code{\link{rx_exec}} automatically register themselves to
be destroyed when the domain within which they were created is closed. This
makes cleanup easier and prevents tasks from one "session" from bleeding into
the next session.
}
\details{
This function can be used to either create and return a domain; or to
evaluate an expression or code block under the domain, and close the domain
when complete.

When passing \code{expr}, this function is reentrant.
}
\examples{

}

