% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{rx_exec}
\alias{rx_exec}
\title{Create a reactive task (observer)}
\usage{
rx_exec(x, env = parent.frame(), quoted = FALSE, label = NULL,
  suspended = FALSE, priority = 0, domain = getDefaultReactiveDomain(),
  autoDestroy = TRUE, ..stacktraceon = TRUE)
}
\description{
Currently just an alias for \code{\link[shiny]{observe}}.
}

