% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{rx_exec_later}
\alias{rx_exec_later}
\title{Run code after a delay}
\usage{
rx_exec_later(delay = 0, expr, env = parent.frame(), quoted = FALSE,
  priority = 0, auto_destroy = TRUE)
}
\description{
Run code after a delay
}
\examples{

rx_session({
  rx_exec_interval(1, {
    print(sample.int(10, 1))
  })

  rx_exec_later(5, {
    rx_stoploop()
  })

  rx_runloop()
})

}

