% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{rx_value}
\alias{rx_value}
\title{Reactive value}
\usage{
rx_value(initial_value = NULL)
}
\arguments{
\item{initial_value}{The starting value of the reactive value object.}
}
\value{
A function. To read the reactive value, call it with no arguments. To
  set the reactive value, call it with the desired value as an argument.
}
\description{
Creates a reactive value object, whose value can be read (reactively) and
set. Setting the value invalidates any reactive conductors or endpoints that
had read the previous value (unless the new value is identical to the
previous value, in which case nothing happens).
}
\examples{
a <- rx_value(10)

r <- rx_calc({
  a()  # read the value
})

a(20)  # set the value
}

